data_pipeline/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.0
yfinance==0.2.33
pandas==2.1.4
beautifulsoup4==4.12.2
feedparser==6.0.10


